requirements.txt
streamlit
pandas
numpy
joblib
torch
monai
matplotlib
shap
google-generativeai
scikit-learn


